chore(sdk-coin-hbar): move account-lib code
- moves hbar related code from account-lib

Ticket: BG-48153
