+++ modules/account-lib/package.json
@@ -48,20 +48,9 @@
     "@bitgo/sdk-core": "^1.0.1",
     "@bitgo/statics": "^6.17.0",
     "@ethereumjs/common": "^2.4.0",
-    "@hashgraph/sdk": "~2.3.0",
-    "@stablelib/hex": "^1.0.0",
-    "@stablelib/sha384": "^1.0.0",
-    "@taquito/local-forging": "6.3.5-beta.0",
-    "@taquito/signer": "6.3.5-beta.0",
-    "@types/lodash": "^4.14.151",
     "bignumber.js": "^9.0.0",
     "bs58": "^4.0.1",
-    "ethereumjs-abi": "^0.6.5",
-    "lodash": "^4.17.15",
-    "long": "^4.0.0",
-    "protobufjs": "^6.8.9",
-    "stellar-sdk": "^10.0.1",
-    "tweetnacl": "^1.0.3"
+    "ethereumjs-abi": "^0.6.5"
   },
   "devDependencies": {
     "@solana/web3.js": "1.31.0",
