+++ modules/sdk-coin-hbar/package.json
@@ -41,7 +41,17 @@
     ]
   },
   "dependencies": {
-    "@bitgo/sdk-core": "^1.0.1"
+    "@bitgo/sdk-core": "^1.0.1",
+    "@bitgo/statics": "^6.17.0",
+    "@hashgraph/proto": "2.0.1-beta.2",
+    "@hashgraph/sdk": "~2.3.0",
+    "@stablelib/sha384": "^1.0.0",
+    "bignumber.js": "^9.0.0",
+    "lodash": "^4.17.15",
+    "long": "^4.0.0",
+    "protobufjs": "^6.8.9",
+    "stellar-sdk": "^10.0.1",
+    "tweetnacl": "^1.0.3"
   },
   "devDependencies": {
     "@bitgo/sdk-api": "^1.0.1",
