+++ modules/sdk-coin-hbar/src/index.ts
@@ -1,2 +1,1 @@
-export * from './hbar';
-export * from './thbar';
+export * from './lib';
