+++ modules/sdk-coin-hbar/src/lib/constants.ts
