+++ modules/sdk-coin-hbar/src/lib/iface.ts
