+++ modules/sdk-coin-hbar/src/lib/keyPair.ts
